+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The library uses [semantic versions](http://semver.org) to identify stable relea
 
 **Features**
 * Implements basic lenses: identity (`lens:id/0`), constant (`lens:const/1`)
-* Implements binary lenses: head of binary (`lens:bh/1`), tail of binary (`lens:bt/`), bit focus (`lens:bf/2`)
+* Implements binary lenses: head of binary (`lens:hbits/1`), tail of binary (`lens:tbits/`), bit focus (`lens:bits/2`)
 
 
 ## Release 4.3.2